requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
joblib>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# API & Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Document Processing
python-docx>=1.0.0

# Presentation (optional)
python-pptx>=0.6.21
